Move `get_time_filter` to settings
